Trigger Images on Base Image Updates
Previously, the image builds would only be triggered by changes to image-
source or ci.  This left out the very common trigger of changes to the
underlying base image of each of the custom Docker images.  Changes to these
images typically include security updated and given that the creation of new
Docker images is completely hands off, there's no reason that these security
updates shouldn't be consumed as soon as they are available.  This change adds
resources for each of the base images and uses them as triggering resources to
the docker image jobs.